Move the cache env in the config file
