+++ package.json
@@ -27,11 +27,9 @@
     "react-test-renderer": "~16.1.1"
   },
   "scripts": {
-    "dev": "next",
-    "dev-cache": "node server.js --port=$PORT",
     "build": "next build",
-    "start": "next start --port=$PORT",
-    "start-cache": "NODE_ENV=production node server.js --port=$PORT"
+    "dev": "node server.js --port=$PORT",
+    "start": "NODE_ENV=production node server.js --port=$PORT"
   },
   "jest": {
     "setupTestFrameworkScriptFile": "<rootDir>/jest.js"
