+++ src/config.dist.js
@@ -1,3 +1,4 @@
-export default {
+module.exports = {
     apiUrl: 'http://your.puzzle.api/api',
+    useCache: true,
 };
