+++ src/services/multiplayerGameService.js
@@ -1,4 +1,4 @@
-import config from '../../config';
+import config from '../config';
 
 const DEFAULT_HEADER_JSON = {
     Accept: 'application/json',
